+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Bake
 
-## Unreleased
+## v2.3.0 - 2026-09-22
 
 ### Added